Move entrypoints into their own directories.
