+++ cmd/server/shared/shared.go
@@ -132,8 +132,8 @@ func Main() {
 		`gitserver: gitserver`,
 		`query-runner: query-runner`,
 		`symbols: symbols`,
-		`lsif-server: node /lsif/out/server.js`,
-		`lsif-worker: node /lsif/out/worker.js`,
+		`lsif-server: node /lsif/out/server/server.js`,
+		`lsif-worker: node /lsif/out/worker/worker.js`,
 		`management-console: management-console`,
 		`searcher: searcher`,
 		`github-proxy: github-proxy`,
