+++ lsif/README.md
@@ -6,7 +6,7 @@ LSIF dumps are generated by running an LSIF indexer in a build or continuous int
 
 ### Architecture
 
-This project is split into two parts, both written in TypeScript. These parts are deployable independently, but for now they run in the same docker container. The HTTP [server](./src/server.ts) receives LSIF uploads and answers LSP queries by looking at a relevant SQLite database on-disk. The [worker](./src/worker.ts) receives job requests via Redis. The worker converts LSIF dump uploads into SQLite databases that can be queried by the server.
+This project is split into two parts, both written in TypeScript. These parts are deployable independently, but for now they run in the same docker container. The HTTP [server](./src/server/server.ts) receives LSIF uploads and answers LSP queries by looking at a relevant SQLite database on-disk. The [worker](./src/worker/worker.ts) receives job requests via Redis. The worker converts LSIF dump uploads into SQLite databases that can be queried by the server.
 
 ### Documentation
