+++ lsif/src/worker/worker.ts
@@ -2,21 +2,21 @@ import * as fs from 'mz/fs'
 import * as path from 'path'
 import express from 'express'
 import promClient from 'prom-client'
-import { convertLsif } from './importer'
-import { dbFilename, ensureDirectory, readEnvInt } from './util'
-import { createLogger } from './logging'
-import { createPostgresConnection } from './connection'
+import { convertLsif } from '../importer'
+import { dbFilename, ensureDirectory, readEnvInt } from '../util'
+import { createLogger } from '../logging'
+import { createPostgresConnection } from '../connection'
 import { Logger } from 'winston'
-import { XrepoDatabase } from './xrepo'
+import { XrepoDatabase } from '../xrepo'
 import { Tracer, FORMAT_TEXT_MAP, Span, followsFrom } from 'opentracing'
-import { createTracer, TracingContext, logAndTraceCall, addTags } from './tracing'
-import { waitForConfiguration, ConfigurationFetcher } from './config'
-import { jobDurationHistogram, jobDurationErrorsCounter } from './worker.metrics'
+import { createTracer, TracingContext, logAndTraceCall, addTags } from '../tracing'
+import { waitForConfiguration, ConfigurationFetcher } from '../config'
+import { jobDurationHistogram, jobDurationErrorsCounter } from '../worker.metrics'
 import { Job, JobStatusClean, Queue } from 'bull'
-import { createQueue } from './queue'
-import { instrument } from './metrics'
-import { purgeOldDumps } from './retention'
-import * as constants from './constants'
+import { createQueue } from '../queue'
+import { instrument } from '../metrics'
+import { purgeOldDumps } from '../retention'
+import * as constants from '../constants'
 
 /**
  * Which port to run the worker metrics server on. Defaults to 3187.
