+++ lsif/supervisord.conf
@@ -8,7 +8,7 @@ command=/kill_supervisor.sh
 events=PROCESS_STATE_EXITED
 
 [program:server]
-command=node /lsif/out/server.js
+command=node /lsif/out/server/server.js
 
 [program:worker]
-command=node /lsif/out/worker.js
+command=node /lsif/out/worker/worker.js
